Update ios-deploy.m
Because the error code was being exported as a signed the "Device Unlock" code was getting sent as -402652958 rather than 3892314338.

Currently all the check_error() macro casts the return to `unsigned int` so also casting the call for image-mounting to do the same; this makes the value come through as an unsigned int which can be compared to the hex values that are expected(e.g. 0xe80000e2 = 3892314338, not -402652958 ).